requirements.txt
flask>=2.3.0
werkzeug>=2.3.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
tensorflow-cpu>=2.13.0
h5py>=3.9.0
gunicorn>=21.2.0
